refactor(core): Distributed test through different files and renamed existing ones to follo the deno naming conventions
