+++ packages/pretch-core/build_fetch_test.ts
@@ -1,9 +1,9 @@
-import { buildFetch } from "@/build-fetch.ts";
+import { buildFetch } from "@/build_fetch.ts";
 import { expect } from "@std/expect";
 import { stub } from "@std/testing/mock";
-import { applyMiddlewares } from "@/middleware/apply-middlewares.ts";
-import { validateStatusMiddleware } from "@/middleware/validate-status.ts";
-import { defaultHeadersMiddleware } from "@/middleware/default-headers.ts";
+import { applyMiddlewares } from "@/middleware/apply_middlewares.ts";
+import { validateStatusMiddleware } from "@/middleware/validate_status.ts";
+import { defaultHeadersMiddleware } from "@/middleware/default_headers.ts";
 import { jwtMiddleware } from "@/middleware/jwt.ts";
 
 type Todo = { userId: number; id: number; title: string; completed: boolean };
